--- a/build.sbt
+++ b/build.sbt
@@ -12,7 +12,6 @@ val compilerOptions = Seq(
     "-unchecked",
     "-deprecation",
     "-Wvalue-discard",
-    "-Wunused:all",
     "-language:strictEquality"
 )
 
@@ -27,7 +26,7 @@ lazy val `kyo-settings` = Seq(
     scalaVersion       := scala3Version,
     crossScalaVersions := List(scala3Version),
     scalacOptions ++= compilerOptions,
-    scalafmtOnCompile := true,
+    scalafmtOnCompile := false,
     organization      := "io.getkyo",
     homepage          := Some(url("https://getkyo.io")),
     licenses          := List("Apache-2.0" -> url("http://www.apache.org/licenses/LICENSE-2.0")),
@@ -45,39 +44,50 @@ lazy val `kyo-settings` = Seq(
     Test / testOptions += Tests.Argument("-oDG"),
     ThisBuild / versionScheme := Some("early-semver"),
     scalacOptions ++= Seq("-release:11"),
+    libraryDependencies += "org.scalatest" %%% "scalatest" % scalaTestVersion % Test,
     Test / javaOptions += "--add-opens=java.base/java.lang=ALL-UNNAMED"
 )
 
-lazy val kyo =
-    crossProject(JVMPlatform)
-        .in(file("."))
-        .settings(
-            name                                   := "kyo",
-            organization                           := "io.getkyo",
-            publishArtifact                        := false,
-            publish / skip                         := true,
-            Compile / packageBin / publishArtifact := false,
-            Compile / packageDoc / publishArtifact := false,
-            Compile / packageSrc / publishArtifact := false,
-            scalaVersion                           := scala3Version,
-            `kyo-settings`
-        ).aggregate(
-            `kyo-scheduler`,
-            `kyo-scheduler-zio`,
-            `kyo-tag`,
-            `kyo-core`,
-            `kyo-direct`,
-            `kyo-stats-registry`,
-            `kyo-stats-otel`,
-            `kyo-cache`,
-            `kyo-sttp`,
-            `kyo-tapir`,
-            `kyo-caliban`,
-            `kyo-bench`,
-            `kyo-test`,
-            `kyo-zio`,
-            `kyo-examples`
-        )
+lazy val kyoJVM = project
+    .in(file("."))
+    .settings(
+        name := "kyoJVM",
+        `kyo-settings`
+    )
+    .aggregate(
+        `kyo-scheduler`.jvm,
+        `kyo-scheduler-zio`.jvm,
+        `kyo-tag`.jvm,
+        `kyo-core`.jvm,
+        `kyo-direct`.jvm,
+        `kyo-stats-registry`.jvm,
+        `kyo-stats-otel`.jvm,
+        `kyo-cache`.jvm,
+        `kyo-sttp`.jvm,
+        `kyo-tapir`.jvm,
+        `kyo-caliban`.jvm,
+        `kyo-bench`.jvm,
+        `kyo-test`.jvm,
+        `kyo-zio`.jvm,
+        `kyo-examples`.jvm
+    )
+
+lazy val kyoJS = project
+    .in(file("js"))
+    .settings(
+        name := "kyoJS",
+        `kyo-settings`
+    )
+    .aggregate(
+        `kyo-scheduler`.js,
+        `kyo-tag`.js,
+        `kyo-core`.js,
+        `kyo-direct`.js,
+        `kyo-stats-registry`.js,
+        `kyo-sttp`.js,
+        `kyo-test`.js,
+        `kyo-zio`.js
+    )
 
 lazy val `kyo-scheduler` =
     crossProject(JSPlatform, JVMPlatform)
@@ -142,14 +152,13 @@ lazy val `kyo-core` =
         .in(file("kyo-core"))
         .settings(
             `kyo-settings`,
-            libraryDependencies += "com.lihaoyi"   %%% "pprint"          % "0.9.0",
-            libraryDependencies += "org.jctools"     % "jctools-core"    % "4.0.3",
-            libraryDependencies += "org.slf4j"       % "slf4j-api"       % "2.0.13",
-            libraryDependencies += "dev.zio"       %%% "zio-laws-laws"   % "1.0.0-RC26"     % Test,
-            libraryDependencies += "dev.zio"       %%% "zio-test-sbt"    % "2.1.1"          % Test,
-            libraryDependencies += "org.scalatest" %%% "scalatest"       % scalaTestVersion % Test,
-            libraryDependencies += "ch.qos.logback"  % "logback-classic" % "1.5.6"          % Test,
-            libraryDependencies += "javassist"       % "javassist"       % "3.12.1.GA"      % Test
+            libraryDependencies += "com.lihaoyi"  %%% "pprint"          % "0.9.0",
+            libraryDependencies += "org.jctools"    % "jctools-core"    % "4.0.3",
+            libraryDependencies += "org.slf4j"      % "slf4j-api"       % "2.0.13",
+            libraryDependencies += "dev.zio"      %%% "zio-laws-laws"   % "1.0.0-RC26" % Test,
+            libraryDependencies += "dev.zio"      %%% "zio-test-sbt"    % "2.1.1"      % Test,
+            libraryDependencies += "ch.qos.logback" % "logback-classic" % "1.5.6"      % Test,
+            libraryDependencies += "javassist"      % "javassist"       % "3.12.1.GA"  % Test
         )
         .jsSettings(`js-settings`)
 
@@ -158,7 +167,7 @@ lazy val `kyo-direct` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-direct"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "com.github.rssh" %%% "dotty-cps-async" % "0.9.21"
@@ -189,7 +198,7 @@ lazy val `kyo-stats-otel` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-stats-otel"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "io.opentelemetry" % "opentelemetry-api"                % "1.38.0",
@@ -202,7 +211,7 @@ lazy val `kyo-cache` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-cache"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "com.github.ben-manes.caffeine" % "caffeine" % "3.1.8"
@@ -213,7 +222,7 @@ lazy val `kyo-os-lib` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-os-lib"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "com.lihaoyi" %% "os-lib" % "0.10.1"
@@ -224,7 +233,7 @@ lazy val `kyo-sttp` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-sttp"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "com.softwaremill.sttp.client3" %%% "core" % "3.9.7"
@@ -236,7 +245,7 @@ lazy val `kyo-tapir` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-tapir"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .dependsOn(`kyo-sttp`)
         .settings(
             `kyo-settings`,
@@ -249,10 +258,10 @@ lazy val `kyo-caliban` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Pure)
         .in(file("kyo-caliban"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .dependsOn(`kyo-tapir`)
         .dependsOn(`kyo-zio`)
-        .dependsOn(`kyo-sttp` % "test->test")
+        .dependsOn(`kyo-sttp`)
         .settings(
             `kyo-settings`,
             libraryDependencies += "com.github.ghostdogpr"       %% "caliban"        % "2.7.0",
@@ -280,12 +289,12 @@ lazy val `kyo-zio` =
         .withoutSuffixFor(JVMPlatform)
         .crossType(CrossType.Full)
         .in(file("kyo-zio"))
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
-            libraryDependencies += "dev.zio" %% "zio"          % zioVersion,
-            libraryDependencies += "dev.zio" %% "zio-test"     % zioVersion,
-            libraryDependencies += "dev.zio" %% "zio-test-sbt" % zioVersion % Test
+            libraryDependencies += "dev.zio" %%% "zio"          % zioVersion,
+            libraryDependencies += "dev.zio" %%% "zio-test"     % zioVersion,
+            libraryDependencies += "dev.zio" %%% "zio-test-sbt" % zioVersion % Test
         ).jsSettings(
             `js-settings`
         )
@@ -298,7 +307,7 @@ lazy val `kyo-examples` =
         .dependsOn(`kyo-tapir`)
         .dependsOn(`kyo-direct`)
         .dependsOn(`kyo-os-lib`)
-        .dependsOn(`kyo-core` % "test->test;compile->compile")
+        .dependsOn(`kyo-core`)
         .settings(
             `kyo-settings`,
             Compile / doc / sources                              := Seq.empty,
@@ -316,6 +325,7 @@ lazy val `kyo-bench` =
         .dependsOn(`kyo-scheduler-zio`)
         .settings(
             `kyo-settings`,
+            Test / testForkedParallel := true,
             // Forks each test suite individually
             Test / testGrouping := {
                 val javaOptionsValue = javaOptions.value.toVector