--- a/kyo-core/shared/src/test/scala/kyoTest/Tagged.scala
+++ b/kyo-core/shared/src/test/scala/kyoTest/Tagged.scala
@@ -4,6 +4,6 @@ import org.scalatest.Tag
 
 object Tagged:
     private def runWhen(cond: => Boolean) = if cond then "" else "org.scalatest.Ignore"
-    object jvmOnly extends Tag(runWhen(Platform.isJVM))
-    object jsOnly  extends Tag(runWhen(Platform.isJS))
+    object jvmOnly extends Tag(runWhen(kyo.internal.Platform.isJVM))
+    object jsOnly  extends Tag(runWhen(kyo.internal.Platform.isJS))
 end Tagged