--- a/kyo-stats-otel/shared/src/test/scala/kyoTest/stats/otel/KyoTest.scala
+++ b/kyo-stats-otel/shared/src/test/scala/kyoTest/stats/otel/KyoTest.scala
@@ -7,10 +7,10 @@ import org.scalatest.freespec.AsyncFreeSpec
 import scala.concurrent.ExecutionContext
 import scala.language.implicitConversions
 
-abstract class KyoTest extends AsyncFreeSpec with BaseKyoTest with NonImplicitAssertions:
+abstract class KyoTest extends AsyncFreeSpec with BaseKyoTest with NonImplicitAssertions {
 
     type Assertion = org.scalatest.compatible.Assertion
     def success = succeed
 
-    override given executionContext: ExecutionContext = kyo.internal.Platform.executionContext
-end KyoTest
+    implicit override def executionContext: ExecutionContext = kyo.internal.Platform.executionContext
+}