--- a/kyo-tapir/shared/src/test/scala/kyoTest/server/NettyKyoServerTest.scala
+++ b/kyo-tapir/shared/src/test/scala/kyoTest/server/NettyKyoServerTest.scala
@@ -2,13 +2,18 @@ package kyoTest.server
 
 import kyo.*
 import kyoTest.KyoTest
+import sttp.tapir.server.netty.NettyConfig
 import sttp.tapir.server.netty.NettyKyoServer
 
 class NettyKyoServerTest extends KyoTest:
 
+    def testServer() =
+        import scala.concurrent.duration.*
+        NettyKyoServer(NettyConfig.default.copy(gracefulShutdownTimeout = Some(5.millis)))
+
     "start stop" in run {
         for
-            bindings <- NettyKyoServer().start()
+            bindings <- testServer().start()
             _        <- bindings.stop()
         yield succeed
     }