--- a/project/plugins.sbt
+++ b/project/plugins.sbt
@@ -9,4 +9,6 @@ addSbtPlugin("org.scalameta" % "sbt-mdoc" % "2.5.2")
 
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "2.0.12")
 
+// addSbtPlugin("com.gradle" % "sbt-develocity" % "1.0.1")
+
 // addSbtPlugin("com.github.sbt" % "sbt-jacoco" % "3.4.0")