- Split root project into `kyoJVM` and `kyoJS`. The JVM project is selected by default, this way the regular development cycle is faster but there's the downside that we might not catch issues in JS before CI. It seems a reasonable tradeoff since it's possible to run the tests manually by selecting the `kyoJS` project.
- Avoid the dependency on `kyo-core` tests in other modules. All other modules end up stuck waiting for the core tests to compile because of the dependency. I had to duplicate some of the code in `KyoTest` into the other modules to achieve that.
- Move `scalafmt` from the build to a commit hook.
- Remove `-Wunused:all` compiler option. It significantly increases the compilation times for tests. We can attempt to add it back once we're on Scala 3.5, which seems to [improve build times](https://github.com/getkyo/kyo/issues/342#issuecomment-2135255841) with the option enabled.

Results with a cold JVM:

**Main**
```
$ sbt ";clean;compile"
[success] Total time: 24 s, completed May 30, 2024, 1:58:13 PM
$ sbt ";clean;test:compile"
[success] Total time: 58 s, completed May 30, 2024, 1:59:58 PM
$ sbt ";clean;test"
[success] Total time: 88 s (01:28), completed May 30, 2024, 2:01:45 PM
```

**This PR (default `kyoJVM` only)**
```
$ sbt ";clean;compile"
[success] Total time: 14 s, completed May 30, 2024, 1:54:29 PM
$ sbt ";clean;test:compile"
[success] Total time: 27 s, completed May 30, 2024, 1:55:51 PM
$ sbt ";clean;test"
[success] Total time: 39 s, completed May 30, 2024, 1:56:59 PM
```

I won't close https://github.com/getkyo/kyo/issues/342 since there's always opportunity to optimize the build.