@fwbrasil 2024-05-31
> LGTM. Perhaps we can migrate to kyo-test for non kyo-core modules...

Good point! Something I noticed is that the tests using zio-test seem to take significantly longer than the ones in scalatest, though.